requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
gradio
requests
python-dotenv